Sensors/requirements.txt
requests
paho-mqtt
gpiozero
adafruit-blinka
adafruit-circuitpython-dht
adafruit-circuitpython-board
